Update patch set 8
Patch Set 8:

Yeah, I´ve been trying to swipe evrything what its swipable and doesn´t work. The three dots for settings are not coming back, so you are not able to disable the full screen mode.

By the way is the fullscreen using the inmersive mode new thing?

Patch-set: 8