OmniTorch: added activity start broadcast
Add new broadcast that starts the torch with activity
and not only the service. This is used e.g. for
starting torch from offscreeen gesture

Added keep screen on settings and fix format of all
sources files to be AOSP conform.

Change fullscreen mode to be immersive only so it can 
be disabled on navbar only devices

Change-Id: I062f322cc3b7c3e2134b6d82f317d1097311c4a2